requirements.txt
openai==0.27.0
pyttsx3==2.90
SpeechRecognition==3.8.1
keyboard==0.13.5
Pillow==9.3.0